Adding a "root" type mapper
Adding a new RootTypeMapperInterface that can be used to map any return or any parameter to a GraphQL output / input type.

Unless the TypeMapperInterface, the RootTypeMapperInterface can map any input/output type (not only objects).
It also has access to the method signature / annotation, etc... and not only the type.