Removing the concept of Hydrator
This PR removes the `HydratorInterface` and all implementations.

When returning an input object from a TypeMapper, the object must now implement the `ResolvableMutableInputInterface` (an input object type that contains its own resolver)
The need for HydratorInterface is gone as there is only one possible correct implementation and it is now in the ArgumentResolver.

Closes #64