Adding type mapper factory
Sometimes, a type mapper requires a the recursive type mapper to be created.
While it is possible to pass it using dependency injection, it is not
easy if you use the simpler SchemaFactory.

Using a TypeMapperFactory, you can now easily create a TypeMapper that
will be passed a recursive type mapper.