Mapping ENUM keys to GraphQL Enum names
So far, ENUM names in GraphQL were mapped to the const values.
This PR changes this to using const keys instead of const values.

So:

```php
class TestEnum extends Enum
{
    private const ON = 'on';
    private const OFF = 'off';
}
```

This will generate an enum with ("ON"/"OFF") instead of ("on"/"off").
This is a more natural match for GraphQL Enum names because enum names are limited to /[_A-Za-z][_0-9A-Za-z]*/ (just like the CONST names and not like the CONST values).