Migrating cache to Symfony Cache contracts
So far, GraphQLite was implementing its own Stampede protection in GlobTypeMapper on top of PSR-16.
Code was hard to read and hard to test.

It turns out that symfony/cache comes with "Cache contracts", a wrapper
built on top of PSR-16 that provides a better API with built-in stampede
protection, and even more (like probabilistic early expiration).

This PR migrates to Symfony cache contracts and allows us to remove a
bunch of code! \o/