Improving non nullable arguments
Currently, if an argument is non-nullable in PHP, it is also non nullable in GraphQL.
But sometimes, it could be interesting for a non-nullable argument in PHP to be nullable in GraphQL.

In the very specific case where we can create the factory with no parameters.
In this case, it makes little sense to force the GraphQL user to pass the parameters.

```php
/**
 * @Factory()
 */
public function createFilter(int $somevalue = 3, string $someothervalue = "foo"): Filter { ... }
```

In this case, if a query looks like this:

```php
/**
 * @Query()
 */
public function myQuery(Filter $filter): ...
```

it makes little sense to force the user to pass a "filter: {}" in the GraphQL parameters. Not specifying the filter should still call the Factory and not trigger an error.

Beware!
If the factory is:

```php
/**
 * @Query()
 */
public function myQuery(?Filter $filter): ...
```

Then we should still aim for "null" as the default value.