Forcing declaration of autowired services through @Autowire annotation
The fact that GraphQLite tries automatically to autowire dependencies leads to problems, especially with autowiring containers.
Indeed, with autowiring containers, it is not immediately clear whether a parameter type-hint means we want an input type or a service from the container.

As a result, a misconfigured factory could go unnoticed as the autowiring container would automatically kick in and respond instead of the input type generator.
This leads to hard to debug code.

A simple solution is to make autowiring explicit. By stating that a parameter is meant to be autowired, we make error messages more relevant and when reading the code, the intent is clearer.