Migrating away from @Parameter annotation
Having a single @Parameter annotation that gathers all the annotations dedicated to a single parameter leads to a weird syntax.

For instance:

```php
/**
 * @Parameter(for="$translator", annotations={
 *    @Autowire
 * })
 */
```

Furthermore, the PHP Annotations extension for PHPStorm deals poorly with nested annotations and we do not have autofill of use statements.

This PR removes the use of nested annotations in Parameter.
The @Parameter is completely removed in favor of 2 annotations:

@UseInputType
and
@Autowire

The signature of the ParameterMapperInterface is modified. Instead of accepting a ParameterAnnotation argument, it now accepts a ParameterAnnotations (notice the s) class that can be used to access any class implementing the ParameterAnnotationInterface.
Both Autowire and @UseInputType now implement the ParameterAnnotationInterface.