Fixing extend type application range
The extended types (applied with @ExtendType) should be applied from the class declared in the annotation, up to the parent class that is also a type.

Until now, the whole class tree was taken in account. As a result, the @ExtendType was applied once for each parent class (when it really needs to be applied once).