Refactoring the wrapping of exceptions
Instead of wrapping exceptions to add location data, they are now modified in place.
In particular, once modified to add location data, they cannot be modified again.
This way, we don't end up with a chain of exceptions where useless location data is added.

Also, this adds a location data for @ExtendType.