Adding a StaticClassListTypeMapper
This is a variant of the GlobTypeMapper that is not performing any glob.
Instead it is being passed a list of classes that are looked into for annotations.
This is useful because the GlobTypeMapper is not looking into packages (only the root package)
and therefore it is important to have a way for packages to provide types too.