Changing the default name given to enum types in the GraphQL schema
Warning! This is a BC break.
Given the fact that enum types where almost non functionals in v4.0.1,
and given the fact it is unlikely they have been used until now, I'm exceptionnaly introducing
this BC break to introduce Enum types that are behaving in a more consistent way.

Enum type name is now the short class name and a @EnumType annotations allows overriding the name if needed.

Closes #232.