Using Native Symfony namespace feature
Before, PSR-16 cache was wrapped in a NamespaceCache (PSR-16 compatible) and then wrapped in a Symfony PSR16Adapter to generate a Symfony cache.

Now, the PSR16Adapter namespace feature is used and NamespaceCache is no more needed when Symfony cache is used.

Note: in the future, GraphQLite will accept PSR-16, PSR-6 or Symfony cache is input.