Add method overloading support
Currently it's not possible to user overloading magic methods.  They're kind of nasty, in general.  However, they're useful in certain cases.  One example where they're particularly useful, is when defining `#[Input]` type classes.  If a `#[Field] property is not `public` you must provide a `setter`.  We could work around this, allowing the `InputType` mapper to set values on `private` and `protected` properties.  However, I'm not really convinced that's the best way to do it.  We should try to avoid side-stepping class intent.

This PR just provides support for `__call` if it's been defined, validating as having a getter or setter.