feat: replace thecodingmachine/class-explorer with kcs/class-finder (#664)
* feat: replace thecodingmachine/class-explorer with kcs/class-finder

Main issue is to let type mappers find types in vendor packages which class-explorer and maintainer is not updating the project.

Symfony and Laravel bundles have to be updated too.
Fixes: #657

* test: check that incorrect classes don't trigger autoloading errors

Thanks @oprypkhantc

* fix: stop caching reflections

* fix: static checks

* fix: stop using cached classes if restoring fails

* test: improve coverage

* fix: prevent possible issue with long-running apps

As each condition applied to finder stays there each place it is applied should be done on cloned instance to avoid accumulation. Main instance of finder is kept with all the conditions provided during configuring. NS is created only from NSFctory where it's cloned so no need to add same clone inside NS.